requirements.txt
pandas==2.0.3
requests==2.31.0
kafka-python==2.0.2
pymssql==2.2.8
python-dotenv==1.0.0
